requirements.txt
discord.py[voice]
wavelink
git+git://github.com/johnwmillr/LyricsGenius.git